[ADF-4461] Moved formbase to appropriate folder
